+++ extension/repo/menu@a6621988-551d-495a-97d8-3c568bca2e9e/router.mjs
@@ -8,8 +8,7 @@
 
 import { web } from '../../api/_.mjs';
 
-let _defaultView = '',
-  $viewRoot;
+let _defaultView = '';
 const _views = new Map();
 
 export function addView(name, loadFunc) {